Robots TXT generator requested changes
