+++ apps/roboshield/src/components/CommonSettings/CommonSettings.tsx
@@ -96,7 +96,7 @@ export default function CommonSettings({
       <SkipToLastStep
         handleSkipToLast={skipToLast}
         lastStep={lastStep}
-        toolTipText={actions?.toolTipText}
+        showRobotsTxt={actions?.showRobotsTxt}
       />
       <StepHint hint={hint} />
       <Box
