--- apps/roboshield/src/components/ExistingRobots/index.ts
@@ -1,3 +1,0 @@
-import ExistingRobots from "./ExistingRobots";
-
-export default ExistingRobots;
