+++ apps/roboshield/src/components/ExistingRobotsTxt/ExistingRobotsTxt.tsx
@@ -101,7 +101,7 @@ export default function ExistingRobots({
       <SkipToLastStep
         handleSkipToLast={skipToLast}
         lastStep={lastStep}
-        toolTipText={actions?.toolTipText}
+        showRobotsTxt={actions?.showRobotsTxt}
       />
       <StepHint hint={hint} />
       <Box sx={{ py: 2 }}>
