+++ apps/roboshield/src/components/ExistingRobotsTxt/index.ts
@@ -1,0 +1,3 @@
+import ExistingRobotsTxt from "./ExistingRobotsTxt";
+
+export default ExistingRobotsTxt;
