--- apps/roboshield/src/components/RobotsGenerator/index.ts
@@ -1,3 +1,0 @@
-import RobotsGenerator from "./RobotsGenerator";
-
-export default RobotsGenerator;
