+++ apps/roboshield/src/components/RobotsTxtGenerator/index.ts
@@ -1,0 +1,3 @@
+import RobotsTxtGenerator from "./RobotsTxtGenerator";
+
+export default RobotsTxtGenerator;
