+++ apps/roboshield/src/components/Sitemaps/Sitemaps.tsx
@@ -45,7 +45,7 @@ export default function Sitemaps({
       <SkipToLastStep
         handleSkipToLast={skipToLast}
         lastStep={lastStep}
-        toolTipText={actions?.toolTipText}
+        showRobotsTxt={actions?.showRobotsTxt}
       />
       <StepHint hint={hint} />
       <Box
